- Added a global symbol table and changed DECORATE parser to put its global   symbols there instead of into AActor. - Changed the expression evaluator's floating point precision to double. - Started rewriting the DECORATE expression evaluator to allow more flexibility.   All the operators use the new functionality but functions, variables and   constants are yet to be redone.   While doing this rewrite I noticed that random2 was always evaluated as const.   This got fixed automatically.
SVN r1264 (trunk)

+++ src/thingdef/thingdef.h
@@ -183,7 +183,7 @@ int ParseExpression (FScanner &sc, bool _not, PClass *cls);
 
 bool IsExpressionConst(int id);
 int EvalExpressionI (int id, AActor *self, const PClass *cls=NULL);
-float EvalExpressionF (int id, AActor *self, const PClass *cls=NULL);
+double EvalExpressionF (int id, AActor *self, const PClass *cls=NULL);
 fixed_t EvalExpressionFix (int id, AActor *self, const PClass *cls=NULL);
